+++ lesson-1/src/components/chatList.js
@@ -1,0 +1,62 @@
+import React from "react"
+import { Input, InputAdornment, ListItem, ListItemText, InputLabel, ListItemSecondaryAction, IconButton } from "@material-ui/core"
+import { Add, Delete } from "@material-ui/icons"
+import { Link } from "react-router-dom";
+
+export class ChatList extends React.Component {
+    constructor(props) {
+        super(props);
+        this.state = {
+            value: "",
+        }
+    }
+
+    render() {
+        return <div style={{ width: "30%", padding: "20px" }}>
+            <div style={{ height: "100%" }}>
+                {
+                    this.props.chats.map((chat, index) => (
+                        <ListItem key={index} style={{}} ContainerComponent="div">
+                            <Link to={`/chats/${chat.id}`}>
+                                <ListItemText primary={chat.name} />
+                            </Link>
+                            <ListItemSecondaryAction>
+                                <IconButton onClick={() => { this.delete(chat.id) }}>
+                                    <Delete />
+                                </IconButton>
+                            </ListItemSecondaryAction>
+                        </ListItem>
+                    ))
+                }
+            </div>
+            <InputLabel htmlFor="chatLabel">Введите название нового чата...</InputLabel>
+            <Input
+                id="chatLabel"
+                value={this.state.value}
+                fullWidth={true}
+                onChange={this.handleChange}
+                endAdornment={
+                    <InputAdornment>
+                        <Add onClick={this.add} />
+                    </InputAdornment>
+                }
+            />
+        </div >
+    }
+
+    handleChange = ({ target }) => {
+        this.setState({ value: target.value });
+    }
+
+    add = () => {
+        if (this.state.value != "") {
+            this.props.addChat(this.state.value);
+            this.state.value = "";
+        }
+    }
+
+    delete = (id) => {
+        this.props.deleteChat(id);
+        if (this.props.chatId == id) this.props.history.push('/index');
+    }
+}
