+++ lesson-1/src/components/index.js
@@ -1,0 +1,3 @@
+export * from "./message";
+export * from "./messageField";
+export * from "./chatList";
