+++ lesson-1/src/store/chat/index.js
@@ -1,0 +1,2 @@
+export * from "./reducer";
+export * from "./actions";
