+++ lesson-1/src/store/chat/types.js
@@ -1,0 +1,4 @@
+export const ADDCHAT = "@chat/add";
+export const ADDMESSAGE = "@message/add";
+export const DELETECHAT = "@chat/delete"
+export const DELETEMESSAGE = "@message/delete"
