+++ lesson-1/src/store/index.js
@@ -1,0 +1,1 @@
+export * from "./create-store";
